minor casting fixes

+++ transport.c
@@ -300,7 +300,7 @@ static void control_loop(mysocket_t sd, context_t *ctx)
 		//read data with stcp_app_recv(sd,dst,size) into dst as a char*
 		//  read data with stcp_app_recv(sd,dst,size) into dst as a char*
 		// Sliding window calculations
-		int max_send_window = std::min(ctx->their_recv_win, (uint16_t)ctx->congestion_win); //CODY: error: no matching function for call to 'min(uint16_t&, tcp_seq&)'
+		int max_send_window = std::min(ctx->their_recv_win, ctx->congestion_win); //CODY: error: no matching function for call to 'min(uint16_t&, tcp_seq&)'
 		int data_in_flight = *(ctx->last_byte_sent) - *(ctx->last_byte_ack);
 		//get data from the app
 		if (stcp_app_recv(sd, ctx->data_buffer, (max_send_window - data_in_flight) - 1) == (size_t)-1){
